+++ lib/active_record_compose/inner_model.rb
@@ -7,48 +7,9 @@ module ActiveRecordCompose
     # @param model [Object] the model instance.
     # @param destroy [Boolean] given true, destroy model.
     # @param destroy [Proc] when proc returning true, destroy model.
-    def initialize(model, destroy: false, context: nil)
+    def initialize(model, destroy: false)
       @model = model
-      @destroy_context_type =
-        if context
-          c = context
-
-          if c.is_a?(Proc)
-            # @type var c: ((^() -> (context)) | (^(_ARLike) -> (context)))
-            if c.arity == 0
-              deprecator.warn(
-                '`:context` will be removed in 0.5.0. Use `:destroy` option instead. ' \
-                'for example, `models.push(model, context: -> { foo? ? :destroy : :save })` ' \
-                'is replaced by `models.push(model, destroy: -> { foo? })`.',
-              )
-
-              # @type var c: ^() -> (context)
-              -> { c.call == :destroy }
-            else
-              deprecator.warn(
-                '`:context` will be removed in 0.5.0. Use `:destroy` option instead. ' \
-                'for example, `models.push(model, context: ->(m) { m.bar? ? :destroy : :save })` ' \
-                'is replaced by `models.push(model, destroy: ->(m) { m.bar? })`.',
-              )
-
-              # @type var c: ^(_ARLike) -> (context)
-              ->(model) { c.call(model) == :destroy }
-            end
-          elsif %i[save destroy].include?(c)
-            deprecator.warn(
-              '`:context` will be removed in 0.5.0. Use `:destroy` option instead. ' \
-              "for example, `models.push(model, context: #{c.inspect})` " \
-              "is replaced by `models.push(model, destroy: #{(c == :destroy).inspect})`.",
-            )
-
-            # @type var c: (:save | :destory)
-            c == :destroy
-          else
-            c
-          end
-        else
-          destroy
-        end
+      @destroy_context_type = destroy
     end
 
     delegate :errors, to: :model
@@ -57,10 +18,10 @@ module ActiveRecordCompose
       d = destroy_context_type
       if d.is_a?(Proc)
         if d.arity == 0
-          # @type var d: ^() -> (bool | context)
+          # @type var d: ^() -> bool
           !!d.call
         else
-          # @type var d: ^(_ARLike) -> (bool | context)
+          # @type var d: ^(_ARLike) -> bool
           !!d.call(model)
         end
       else
@@ -106,13 +67,5 @@ module ActiveRecordCompose
     private
 
     attr_reader :model, :destroy_context_type
-
-    def deprecator
-      if ActiveRecord.respond_to?(:deprecator)
-        ActiveRecord.deprecator
-      else # for rails 7.0.x or lower
-        ActiveSupport::Deprecation
-      end
-    end
   end
 end