+++ spec/support/form.rb
@@ -93,61 +93,6 @@ class ComposedModelWithConditionalDestroyContextWithMethodName < ActiveRecordCom
   def blank_profile? = firstname.blank? && lastname.blank? && age.blank?
 end
 
-class ComposedModelWithDestroyContextByDeprecatedOption < ActiveRecordCompose::Model
-  def initialize(account, attributes = {})
-    @account = account
-    @profile = account.profile || account.build_profile
-    super(attributes)
-    models.push(account)
-    models.push(profile, context: :destroy)
-  end
-
-  delegate_attribute :name, :email, to: :account
-  delegate_attribute :firstname, :lastname, :age, to: :profile
-
-  private
-
-  attr_reader :account, :profile
-end
-
-class ComposedModelWithConditionalDestroyContextByDeprecatedOption < ActiveRecordCompose::Model
-  def initialize(account, attributes = {})
-    @account = account
-    @profile = account.profile || account.build_profile
-    super(attributes)
-    models.push(account)
-
-    context = ->(p) { p.firstname.blank? && p.lastname.blank? && p.age.blank? ? :destroy : :save }
-    models.push(profile, context:)
-  end
-
-  delegate_attribute :name, :email, to: :account
-  delegate_attribute :firstname, :lastname, :age, to: :profile
-
-  private
-
-  attr_reader :account, :profile
-end
-
-class ComposedModelWithConditionalDestroyContextWithNoBlockArgumentByDeprecatedOption < ActiveRecordCompose::Model
-  def initialize(account, attributes = {})
-    @account = account
-    @profile = account.profile || account.build_profile
-    super(attributes)
-    models.push(account)
-    models.push(profile, context: -> { blank_profile? ? :destroy : :save })
-  end
-
-  delegate_attribute :name, :email, to: :account
-  delegate_attribute :firstname, :lastname, :age, to: :profile
-
-  private
-
-  attr_reader :account, :profile
-
-  def blank_profile? = firstname.blank? && lastname.blank? && age.blank?
-end
-
 class CallbackOrder < ActiveRecordCompose::Model
   def initialize(tracer)
     @tracer = tracer